+++ packages/signers/raw-tx-creator/src/index.ts
@@ -1,2 +1,3 @@
 export * from "./from-raw-signer"
+export * from "./v5-signer"
 export type { TxCreator, SignerTxCreator } from "@polkadot-api/tx-creator"